operator: Refactor instance reconcilers - extract `refactorAfter` to common - move registry auth to artifact reconciler - run pre-flight checks before building the distro
